requirement.txt
pandas
python-dotenv
openpyxl
snowflake-connector-python[pandas]
